Moved memory-cached and disk-cached discovery to their own packages
