+++ staging/src/k8s.io/client-go/discovery/cached/disk/round_tripper.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package discovery
+package disk
 
 import (
 	"net/http"
