+++ staging/src/k8s.io/client-go/discovery/cached/memory/BUILD
